pyproject.toml
[tool.poetry]
name = "hdr"
version = "0.1.0"
description = ""
authors = ["Zzznorlax <legohulk27@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.5.5"
Pillow = "^9.0.1"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
